Made the add category system more robust and automated.
Finalized the category deleter.
ToDo: Make a check to make sure that the item being deleted is not a parent.

Laid groundwork for the category editor.

Tweaked database for future additions. Reformatted database creation for easier understanding.
Hardcoded parent to be 0 instead of null for data info that ws not supplied.
Added a none selector to parent spinner to allow for no parent to be selected.
Finalized the delete by id class to remove data from the database properly.