Complete rebuild of the database helper class, complete with comments.
The class is now completely generic. which means that it will reduce the complexity of the project and cut down on unnecessary coding.
Database is now closed after every operation. this optimizes the program and potentially prevents issues down the road.
Comments have been added to allow for easier understanding of the class and methods.
Full functionality for adding, deleting, editing and gathering information has been implemented.
Standardized returns and names of all methods for easier coding.